requirements.txt
asgiref==3.6.0
Django==4.1.7
gunicorn==20.1.0
python-creditcard==0.0.1
sqlparse==0.4.3
